try to fix circular import issue
